Moved CanGenerateProfileWithEntitlements to projectmanager
